Fix Copy Link menu
